Realocando controller e rota de /produtos
