+++ src/routes/productsRouter.js
@@ -1,8 +1,10 @@
 import { Router } from "express";
-import { PaginaProduto } from "../controllers/productsController.js";
+import { PaginaProduto, produtos } from "../controllers/productsController.js";
 
 const productsRouter = Router();
 
 productsRouter.get("/produtos/:id", PaginaProduto);
+authRouter.get("/produtos", produtos);
+
 
 export default productsRouter;
